Final Commit
Ready for review. Renamed files to their proper name. Added in variables for the time left. A function is ran every second to update a counter. The counter is updated by finding the element using the id. Adds in a function to stop the clock from running every second. Activated strict mode.